requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
transformers
pillow
mss
accelerate
huggingface_hub